refactor: move Dashboard functions and useState variables to a custom hook
